Move Linux-specific ifname logic to tun_linux.go and place a generic in tun_other.go
