+++ src/yggdrasil/tun.go
@@ -18,16 +18,6 @@ func (tun *tunDevice) init(core *Core) {
   tun.core = core
 }
 
-func (tun *tunDevice) setup(ifname string, addr string, mtu int) error {
-        config := water.Config{ DeviceType: water.TUN }
-        if ifname != "" && ifname != "auto" { config.Name = ifname }
-	iface, err := water.New(config)
-	if err != nil { panic(err) }
-  tun.iface = iface
-  tun.mtu = mtu //1280 // Lets default to the smallest thing allowed for now
-  return tun.setupAddress(addr)
-}
-
 func (tun *tunDevice) write() error {
   for {
     data := <-tun.recv
@@ -55,4 +45,3 @@ func (tun *tunDevice) read() error {
 func (tun *tunDevice) close() error {
   return tun.iface.Close()
 }
-
